Use broadcast rules on 0.6 (fixes dot-operators depwarn)
This is a breaking change, because on Julia 0.6 the array eltype no longer "wins": the output is the same as what you'd get from adding the scalars.
